Issue #11605: Code refactoring
Test content creation methods in a single place.
